feat(18228): move all content from views/Pricing
